RSI Bot/requirements.txt
requests==2.31.0
requests-oauthlib==1.3.1
python-dotenv==1.0.0
schedule==1.2.0